Feature: Bankid Lambda Integration
Earlier the communication towards bankid was made thorugh a nodejs microservice called authentication, which had the responsiblitiy for all authentication methods

Now since we are porting functionallity to AWS Lambdas with the serverless framework, bankid will be an independent collection of lambda operations and we move the routing responsiblity to the AWS gateway instead.

The changes made in this repo makes it possible to access the bankid auth, sign, collect and cancel actions thorugh lambdas.