feat(token generation): replacing json payload with authorization code grant
Before generating a token was based on passing a json payload with credentials. This behavior is now
replaced with the Oauth Authorization Code Grant. This new implementation utalises passing of query
params for obtaining a token. Currently. Instead of passing values in the json body we pass them as
a signed jwt in a query param.

BREAKING CHANGE: Passing personalNumber in the json payload of the post request have been replaced
with the passing of a jwt token in the query string params. The BankID collect should from this
point forward return a authorization code signed with a secret stored in AWS secrets manager.