refactor(config): auth key have been added with related information
The config for token secrets config defintition was brekaing the per service strucutre in the config
object. Therefore a auth key have been added to the config where all config values that relates to
the auth service should be placed.

BREAKING CHANGE: Secrets have been renamed in the auth config, to better represent the acctual
usage.