feat(generatetoken): passing arguments in the request body instead of query params
Passing arguments in query string params is not a valid implementation of the OAuth specification.
The arguments should instead be passed in the request body. The request params is therefore passed
as a json payload in the request body

BREAKING CHANGE: The query string params is replaced and from now on a json body should be defined
in the post request.