feat(createvivacase): new lambda for creating a case with viva as provider for a application period
This new lambda createVivaCase is triggered to run when an bankid.collect event is sent through the
eventBridge. The lmabda iteself does two things, first it checks if a user is allowed to apply for
the coming period and second it creates a case in the dynamodb cases table with period data retrived
from an viva application.