fix(viva-ms): event pattern for triggering the workflow lambda have been updated
The lambda syncWorkflow is triggerd by an event that fires when the status is changed on a case in
DynamoDB. Earlier the pattern on the event rule was checking the string value of case statuses.
However the case status have been updated is now an object. So the event rule have been updated to
look on the type key in a status object instead.