feat(viva-ms): new lambda for submitting a completion case to viva adapter
These changes adds support for building and sending a application completion request to the viva
adapter. The lambda responsible for sending this request parses the answer data that belongs to a
completion form and creates the request body. The main parsing logics is to retrive images from s3
based on information in the answer data and categories the images into attachment objects that the
completions endpoint in the viva adapter expects.