fix(viva-ms): changed prefix on status on the eventBridge pattern
The status that should trigger the lambda was set to active:completionSubmitted:viva. This would
never work with the rest of the event flow. So intstead the lambda is now triggering when the status
type on a case is active:submitted:viva. Since the submitApplication lambda have the same trigger
the lambda also checks if the currentFormId is the completionFormId present in the ssm params for
viva cases