fix(submitcompletion): setting attachment name to uploadedFilename instead of filename
The filename attribute that is passed from the client to an item in DynamoDB Cases Table, was not
always set due to limit in the image library that is used. Since the viva adapter requires a name to
be passed on an attachment it would break. To fix this issue we now use the name of the uploaded
file in s3 instead, since that value will never be null