feat(create-viva-case): setting expirationTime on viva case creation
Earlier no expirationTime was set on a newly created viva case. This caused the case to presist in
the database if no updates to it was made. To fix this issue a initial expirationTime have been set
to 3 days (72 hours), removing an case that is notstarted is no issue since a new case will be
created if the period is open in viva.