refactor(create-viva-case): refactor extraction of personalNumber from PK
Earlier the we retrived the personalNumber from the PK by using the JS replace function. This
behavior is however not consitent with the overall codebase where we use the JS funtion substring
instead. Therefore the syntax have been updated here to use the substring syntax in order to keep
the codebase more consistent