feat: 🎸 add logic to inject period date into answer list
This implementation is ugly as hell but it's good enough for now. The
generated PDF should display the viva application period date at the top
right corner on the first page. This is done by grabbing the period
attribute from the submitted case and formating it in a human readable
format. Then add this date string to a machine made answer object. Last
step is to add this newly created answer object to the answer list on
the submitted case.