feat(update-case): set status for case based on patterns in the request data
Earlier we allowed the setting of a status on a case by passing a value on the attribute statusType
in the json paylod, this behavior is now deprected. Now instead the lambda looks on the data in the
json payload to figure out if the case status should be updated and what that status is. The lambda
can currently set three statuses on a case active:ongoing, active:signature:completed,
active:submitted based on the answers attribute and a new attribute called signature from the json
payload.

BREAKING CHANGE: All clients that uses the /cases with the PUT method will need to update their
requests in order to comply with these new changes.