fix(update-case): checking for encrypted fails due to variations in data types
The areEncryptedAnswers function is responsible for checking if the answers data is encrypted or
not. This function exptected an object to be passed to validate if the key encryptedAnswers was
present, however when answers are encrypted the answers data is not longer an object but instead an
array of answers. this caused the function to fail. Therefore an extra check have been added to
check if the data is an array or not, if the answers data is an array the function will return false
indicating that the answers are not encrypted but decrypted.