perf(update-case): improved validation with joi validation
Added Joi validation to validate the incoming request data. This validation checks for the
following: is there a currentFormId present, is there an currentPosition object, is there an answers
attribute (object or array) and is there a signature object. All these attributes will be validated
and if the validation fails a BadRequestError 400 will be returned to the client