fix(update-case): fixed syntax errors for imports and usage of personalNumber
Some of the imports in the updateCase lambda was not correct and would therefore cause problems with
the validation of request data. Along side of this the destructing of personalNumber was made after
the first usage upon retrival of the user case, this hade to be moved further up in the function.