feat: 🎸 add helper functions
getUserByRole is used to find a user by specifc role, in this case
coApplicant to be set as GSI1 index. getCasePersonList generates a list
of person related to the viva case, like client, partner and child. The
list is later used to make it possible for more than one to sign the
case before submitting to viva