fix: 🐛 event rule pattern viva-ms
The updateCase lambda sets an general status type of active:submitted
without the use of provide suffix. This is to not to be bound to a
specific provider by the status type. The submitApplication and
submitCompletion lambdas would not trigger if status type includes the
suffix 'viva' when using DynamoDb filter expression 'begins_with'.
Instead the event pattern rule is updated to include the attribute
'provider' to ensure that the case provider is 'viva'