hotfix: remove Limit statement
Limit 1 instructs DynamoDB to return only one item when querying the table. The filter expression will then filter only on that one item even if there exists many cases for the same user. The correct way to query is to query on all cases based on the key condition. The filter expression will then filter on the entire query result.