fix(bankid-collect): if user-agent is MittHelsingborg, put event and create authorizationCode
Added support for checking if the incoming request user-agent is from our MittHelsingborg client. If
it is we put an bankid.collect event and create an authorizationCode to be used for authentication.
If it's any other client that sends an request towards the bankid collect endpoint we return the
bankid response and nothing more, no events no authorizationCode creation.

BREAKING CHANGE: New behaviour for the bankid collect endpoint, any requests made without an
user-agent header containing MittHelsingborg will not trigger an event to navet.poll