fix: remove duplicate caching of ssm parameters
because we added caching of ssm parameters in params.read function,
there is no need to also cache them in booking.js file. We can instead
rely that params.read is caching them properly. This means that we
can remove the global values in the booking.js file.
Did also remove the getSsmParameters wrapper function for params.read.