feat(cases-ms): a new service for non api related case lambdas
A new service have been created in for non api related case related lambdas. The first lambda in
this service is used to update pdf attributes on a case when a pdf generate event have been put on
the aws eventbridge. The lambda retrives an pdf from s3, gets the related case from dynamodb and
updates the pdf attributes on the retrived case.