refactor(pdf-ms/generate): moved the logic for generating pdf from html to a separate function
The html to pdf logic have been moved to a separate function in order to contain the luanch and
closing of the chromium browser. Before the chromium browser was not closed if an error occured
further up in the main function of the lambda. This is no longer the case, we start the browser and
close it within a defined scope to keep that logic clean