feat(generaterecurringcasehtml): new lambda for generating html from case with handlebar template
This new lambda generates html from case data with handlebars. The lambda triggers on a dynamodb
events that fires on case modifications where the case provider is VIVA and the status is set to
active:submitted. The lambda then retrives a handlebars template file from an s3 bucket, converts
case data to a data-strucutre that we can work with in the handlebar template, passes the data into
the template and generates html. Finally it saves the html to s3 and outputs an event on the
eventbridge