refactor(viva-ms): added support for retriving multiple notes from answers
Before the creation of notes in the template helper only supported the retrival of one field based
on a hardcoded id. This behavior has no been changed in order to support retrival of multiple notes
answers. This is done by tagging each answer with the tags notes and note. We simply filter out
answers that contains these tags and then create a note object for each answer and return them as an
array of notes in the template data.