feat: 🎸 (viva-ms) add put event
Will put event on two conditions. (A) Are viva status codes of type
completions then we're telling the world just that. (B) If not we tell
the world that the check executed well. In this way we can have two
different lambdas take actions on A or B where B is in this case to
check if completions due date has passed