Fix/completions duedate refined (#374)
* fix: (viva-ms) completions due date

* feat: (viva-ms) add lambda syncing completions

* fix: 🐛 (viva-ms) get 'status' attribute from event detail

The lambda applicationStatus is putting viva status code via 'status'
event detail attribute

* feat: 🎸 (viva-ms) add put event

Will put event on two conditions. (A) Are viva status codes of type
completions then we're telling the world just that. (B) If not we tell
the world that the check executed well. In this way we can have two
different lambdas take actions on A or B where B is in this case to
check if completions due date has passed

* refactor: 💡 (viva-ms) remove obsoleted code

The lambda - syncCaseCompletions handles operation regarding updating
completions on the case. This lambda handles updating case status,
state, select correct completion form and resets applicant signature.
This based on the current completion data.

* refactor: 💡 (viva-ms) to make tests executable

* feat: 🎸 (constants) add new state

Will be set when completion attachements is uploaded to Viva and waiting
for Viva-admin to verify.

* test: 💍 (viva-ms completions) add tests

Testing helper function in completions.js. Used by lambda
setCaseCompletions.js to figure out status, state etc by looking on the
completions data

* feat: 🎸 (viva-ms completions) add helper functions

* fix: 🐛 (completions) typo in function name

* feat: 🎸 (completions) add generic status type contants

* test: 💍 (cases-api status type) add tests

* fix: 🐛 (completions helper) export function syntax

* fix: 🐛 (viva-ms due date) return case id on update

* fix: 🐛 (viva-ms) remove pattern in event rule

Will be implemented in a future release together with an overhaul of the
completions flow

* fix: 🐛 (viva-ms) falsy validation implementation

* fix: 🐛 (viva-ms completions) event rule pattern

* refactor: 💡 (viva-ms) is due date expired

* refactor: 💡 (viva-ms) get locale date

* refactor: 💡 (viva-ms) due date args name

The value passed to the function is milliseconds

* refactor: 💡 (viva-ms) undo destructing

* style: 💄 (viva-ms) group import better

* refactor: 💡 (viva-ms) revert extracting to functions

* refactor: 💡 (viva-ms tests) follow best practices