fix: 🐛 (users-ms) users with no address to be created
findUser lambda is triggered both by navet-ms and cases-api. When triggered by cases-api event detail is only including the personal number from the jwt token in the header.

If user not exisits in the users table this will result in the createUser lambda creating the user without any personal details like adress, civilStatus etc due to this info is not provided in the event body.

This commit will fix it by checking if userDetails includes personalName, firstName and lastName and then
create the user otherwise createUser lambda will just ignore it